Fixed SoapUI test and problems with my sql connector + build version
The SoapUI test can now run automatically + added a small load test.
Sql connector downgraded mysql-connector-java to a version that matches
DB version(Things worked before but with a warning)

Reverted log4j.properties because a "post" start script is executed by
Tomcat script tries to apply patch on these properties to update level
of logging. The patch is only present in "Devel" and system works fine
but the script cant still apply the patch. Since this only concerns "devel"
we keep this for now, and create another ticket regarding solving this.

Minor changes in the Pom.xml to make the build produce a properly named
war file